move file out of postgres models dir so db starts
